Add base color scheme to map
This was not in the original color scheme, but it's in the 16-color
versions so I'm guessing the assumption is you set these in the
terminal emulator.

I think it makes more sense to include them here, since nowadays
terminal emulators can display an arbitrary amount of colors, and it's
conceivable that someone could want different schemes in Neovim and the
rest of the terminal.