Refactor
`legacy.lua` should be more or less 1:1 with the original color scheme,
if it's not please open an issue or send a PR. There is currently no
point in using this over the original vim version, but there is room to
add Treesitter and LSP groups that are consistent with the original
color scheme.

I will be adding Treesitter and LSP support, as well as support for
other plugins (that I use). I will however be diverging slightly from
the original color scheme, but the legacy version will always remain
untouched. I obviously like this color scheme, so it won't be anything
massive, mostly just taking advantage of the added granularity
Treesitter and LSP allow for.

Feedback is very welcome, just open an issue.